now listens on named pipe for commands, adds script to communicate with boss

+++ src/boss/Boss.cpp
@@ -6,6 +6,10 @@
 
 #include <iostream>
 
+#define MAN_RESTART 'r'
+#define MAN_KILL    'k'
+#define MAN_START   's'
+
 namespace boss {
 
 Boss::Boss(boost::shared_ptr<AL::ALBroker> broker_, const std::string &name) :
@@ -17,7 +21,9 @@ Boss::Boss(boost::shared_ptr<AL::ALBroker> broker_, const std::string &name) :
     led(broker),
     manPID(-1),
     manRunning(false),
-    shared(NULL)
+    shared_fd(-1),
+    shared(NULL),
+    fifo_fd(-1)
 {
     std::cout << "Boss Constructor" << std::endl;
 
@@ -42,9 +48,17 @@ Boss::Boss(boost::shared_ptr<AL::ALBroker> broker_, const std::string &name) :
         std::cout << "Tried to bind postprocess, but failed, because " + e.toString() << std::endl;
     }
 
+    fifo_fd = open("/home/nao/nbites/nbitesFIFO", O_RDONLY | O_NONBLOCK);
+    if (fifo_fd <= 0) {
+        std::cout << "FIFO ERROR" << std::endl;
+    }
+
     std::cout << "Boss Constructed successfully!" << std::endl;
 
     startMan();
+
+    // This will not return.
+    listener();
 }
 
 Boss::~Boss()
@@ -56,11 +70,24 @@ Boss::~Boss()
     // Close shared memory
     munmap(shared, sizeof(SharedData));
     close(shared_fd);
-    sem_close(semaphore);
+}
+
+void Boss::listener()
+{
+    while(1)
+    {
+        checkFIFO();
+        sleep(2);
+    }
 }
 
 int Boss::startMan() {
     // TODO make sure man isn't running yet
+    if (manRunning) {
+        std::cout << "Man is already running. Will not start." << std::endl;
+        return -1;
+    }
+
     std::cout << "Building man!" << std::endl;
     pid_t child = fork();
     if (child > 0) {
@@ -73,14 +100,21 @@ int Boss::startMan() {
     else {
         std::cout << "COULD NOT DETACH MAN" << std::endl;
         manRunning = false;
+        return -1;
     }
-    std::cout << "\t\t\tMan built!" << std::endl;
+    std::cout << "\tMan built!" << std::endl;
+    return 1;
 }
 
 int Boss::killMan() {
     // TODO make sure man is actually running
+    if (!manRunning) {
+        std::cout << "BOSS: Man is not running. Cannot kill" << std::endl;
+        return -1;
+    }
 
     kill(manPID, SIGTERM);
+    manRunning = false;
     return 0; // TODO actually return something
 }
 
@@ -128,6 +162,7 @@ int Boss::constructSharedMem()
 
 void Boss::DCMPreProcessCallback()
 {
+    //std::cout << "Pre process" << std::endl;
     std::string joints;
     std::string stiffs;
     std::string leds;
@@ -139,36 +174,32 @@ void Boss::DCMPreProcessCallback()
     if (index != -1) {
         Deserialize des(shared->command[index]);
         des.parse();
-        std::cout << des.nObjects() <<" objects in this many bytes: " << des.totalSize() << std::endl;
         joints = des.stringNext();
-        std::cout << "joints" << std::endl;
         stiffs = des.stringNext();
-        std::cout << "stiffs" << std::endl;
         leds = des.string();
-        std::cout << "leds" << std::endl;
         shared->commandReadIndex = des.dataIndex();
-        std::cout << "Read 'em" << std::endl;
     }
     // End sem
     if (index == -1) return;
-    return;
 
     JointCommand results;
+    //std::cout << "Boss joints" << std::endl;
     results.jointsCommand.ParseFromString(joints);
-    std::cout << results.jointsCommand.DebugString() << std::endl;
+    //std::cout << results.jointsCommand.DebugString() << std::endl;
     results.stiffnessCommand.ParseFromString(stiffs);
-    std::cout << results.stiffnessCommand.DebugString() << std::endl;
+    //std::cout << results.stiffnessCommand.DebugString() << std::endl;
     messages::LedCommand ledResults;
     ledResults.ParseFromString(leds);
-    std::cout << ledResults.DebugString() << std::endl;
+    //std::cout << ledResults.DebugString() << std::endl;
+    //std::cout << "Boss joints done " << std::endl;
 
-    //enactor.command(angles.jointsCommand, angles.stiffnessCommand);
-    //led.setLeds(leds);
+    //enactor.command(results.jointsCommand, results.stiffnessCommand);
+    led.setLeds(ledResults);
 }
 
 void Boss::DCMPostProcessCallback()
 {
-    std::cout << "Post process!" << std::endl;
+    //std::cout << "Post process!" << std::endl;
     SensorValues values = sensor.getSensors();
 
     std::vector<SerializableBase*> objects = {
@@ -193,5 +224,31 @@ void Boss::DCMPostProcessCallback()
     // End Semaphore here! ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~``
 
     //std::cout << "We used this much space for sensors " << usedSpace << std::endl;
+}
+
+void Boss::checkFIFO() {
+    char command[2];
+
+    size_t amt = read(fifo_fd, &command, 2);
+
+    if (amt == 0) {
+        return; // Read nothing
+    }
+
+    switch(command[0]) {
+    case MAN_RESTART:
+        std::cout << "MAN_RESTART" << std::endl;
+        killMan();
+        startMan();
+        break;
+    case MAN_KILL:
+        std::cout << "MAN_KILL" << std::endl;
+        killMan();
+        break;
+    case MAN_START:
+        std::cout << "MAN_START" << std::endl;
+        startMan();
+        break;
+    }
 }
 }
